2nd year/sdi/VideoGamesAPI/assignment1SDI/requirements.txt
gunicorn
django==4.1.7
requests==2.28.1
ptvsd==4.3.2 # Required for debugging.

djangorestframework~=3.14.0
google-cloud-ndb